requirements.txt
gradio>=4.44.0
PyYAML>=6.0.1
requests>=2.31.0
playwright>=1.46.0
mem0ai>=0.1.100
